Added KYC methods
